backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "inferencesh>=0.6.19",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
